backend/requirements.txt
django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
python-dateutil==2.8.2
pillow==10.1.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0 